pyproject.toml
[project]
name = "toio-bc"
version = "0.1.0"
description = "Behavior cloning framework for Sony toio Core Cube robot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "websockets>=11",
    "pyyaml",
    "bleak",
    "numpy",
    "pillow",
    "torch>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/toio_bc"]
